r2-fix: split SHA gate logic to sha_gate.rs (file_size_invariant fix)
Extract check_sha_gate, extract_pr_number, fetch_pr_head_sha and
their 7 tests from comms.rs to new sha_gate.rs module.

comms.rs: 805 → 657 LOC (< MAX_LOC=700)
sha_gate.rs: 149 LOC (new file, < MAX_LOC=700)

file_size_invariant test now passes.
